pyproject.toml
[tool.poetry]
name = "edu-data-modeler"
version = "0.1.0"
description = ""
authors = ["MalshaDissanayake <dmkmd333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^28.4.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
